requirements.txt
dnspython==2.7.0
Faker==35.2.0
pymongo==4.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
typing_extensions==4.12.2
